// src/app/core/models/user-profile/nUIDValidationStatus.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * The entered CUI (National Unique Id) during the registration of a company must be checked against Recom Database
 */
export type NUIDValidationStatus = 'NotYetValid' | 'Invalid' | 'Valid';

export const NUIDValidationStatus = {
    NotYetValid: 'NotYetValid' as NUIDValidationStatus,
    Invalid: 'Invalid' as NUIDValidationStatus,
    Valid: 'Valid' as NUIDValidationStatus
};