// src/app/core/models/user-profile/userCustomerMappingOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CustomerOutputModel } from './customerOutputModel';
import { UserCustomerMappingRole } from './userCustomerMappingRole';
import { UserCustomerMappingStatus } from './userCustomerMappingStatus';

/**
 * Each user can have one or more customers associated with the account  The association is represented by the UserCustomerMapping
 */
export interface UserCustomerMappingOutputModel { 
    /**
     * The type of mapping between the user and the customer
     */
    mappingRole?: UserCustomerMappingRole;
    /**
     * The status of the mapping from Omro perspective
     */
    mappingStatus?: UserCustomerMappingStatus;
    /**
     * The customer detail
     */
    customer?: CustomerOutputModel;
    /**
     * If false, the user cannot select the company. It is shown only for a short period of time before deletion.
     */
    mappingEnabled?: boolean;
}